refactored, housekeeping
